fix: js파일 역할에 따라 나눠서 폴더링
