+++ week2/constants/shoppingList.js
@@ -10,7 +10,7 @@ export const SHOPPING_LIST = [
     id: 2,
     category: "food",
     image: "./assets/images/쌀국수.jpeg",
-    title: "곱창",
+    title: "쌀국수",
     price: "12000",
   },
   {
@@ -55,57 +55,51 @@ export const SHOPPING_LIST = [
     title: "알리오올리오",
     price: "11000",
   },
+
   {
     id: 9,
-    category: "food",
-    image: "./assets/images/알리오올리오.jpeg",
-    title: "알리오올리오",
-    price: "11000",
-  },
-  {
-    id: 10,
     category: "food",
     image: "./assets/images/수플레.jpeg",
     title: "수플레",
     price: "10000",
   },
   {
-    id: 11,
+    id: 10,
     category: "place",
     image: "./assets/images/한강.jpeg",
     title: "한강",
     price: "0",
   },
   {
-    id: 12,
+    id: 11,
     category: "place",
     image: "./assets/images/놀이공원.jpeg",
     title: "놀이공원",
     price: "50000",
   },
   {
-    id: 13,
+    id: 12,
     category: "place",
     image: "./assets/images/스키장.jpeg",
     title: "스키장",
     price: "100000",
   },
   {
-    id: 14,
+    id: 13,
     category: "place",
     image: "./assets/images/야구장.jpeg",
     title: "야구장",
     price: "13000",
   },
   {
-    id: 15,
+    id: 14,
     category: "activity",
     image: "./assets/images/네컷사진.jpeg",
     title: "네컷사진",
     price: "5000",
   },
   {
-    id: 16,
+    id: 15,
     category: "activity",
     image: "./assets/images/카공.jpeg",
     title: "카페에서 공부",
