+++ week2/page/main/navigateToMain.js
@@ -1,9 +1,8 @@
 const select = (selector) => document.querySelector(selector);
-const selectAll = (selector) => document.querySelectorAll(selector);
 
 // 메인 페이지 이동
-const homeBtn = select(".navigate_home_btn");
-const navigateHome = () => {
+export const homeBtn = select(".navigate_home_btn");
+export const navigateHome = () => {
   location.href = "index.html";
 };
 homeBtn.addEventListener("click", () => {
